Stub out new setup for determining names
This is the generic framework I will work in. We keep the old method on
the RNA model (det_description) with the same signature however I move
code around. To begin with, the old logic is split by the parts that
recompute data vs those that do not. The part that looks up a stored
description remains in the model and should have the same behavior of
use precomputed if it exists, otherwise compute one.

The recomputing logic is moved into it's own module,
rnacentral.portal.utils.descriptions. This calls a single function which
will use the new method if possible otherwise it will use the old
method. As the new method is expanded to cover more types of RNA and
databases the old method will be used less. Eventually we will just
remove the old method entirely.

Currently, the new method doesn't actually work as this is just the
framework of what will happen. It does show the general features of how
determining the new descriptions will work however. First we determine
the RNA type (ignoring the species constraint) then we select a database
to use the naming from. We have an ordered choice among databases to
select the one we believe gives the best names.