First set of changes for #101
This begins the work on using a better method for selecting names for a
RNA's. This does appear to fix some of the issues in #101. It will
reject annotations of miRNA for things that are too long. Those will be
called precursors now. In addition, it will select a better name for
miRNA/precursors now. It has an ordered choice among databases and then
if there are several entries from the best database it picks the
longest. This does not fix many of the issues, nor does it store the
data which is required for improving search, but it is the start of
those projects.