Add function for computing the rna_type
Computing the rna_type is a key part of finding a good description for a
sequence. It is also generally useful for searching. Thus I add a
function to compute the rna_type. This function uses the new rule based
approach for determining the rna_type only. It is possible for this
approach to return a set of more than one rna_type (due to issues in
sequence labels that are not corrected yet) or an empty set. The empty
set should not happen however this has not yet been tested completely.