Add some tests for naming
These test probably don't run yet but exist for me to outline the
behaviors I want. The idea is to get a bunch of examples of naming that
I like and then ensure the code matches it. So far this just comes from
me browsing the site be eventually we should include examples from
people who found issues.