Add utility method to get the valid xrefs
The currently valid xrefs are those that are not deleted and match the
species constraint (if any constraint), if there are no such xrefs then
all xrefs should be used. I added a function to find such xrefs and call
it where needed. This has been modified from code in the score_method
module. The function is likely useful enough to be added to the model,
but I'm not sure.