Add custom test runner
In our setup we already have a large existing database. This cannot be
created easily nor can we drop it. We also don't use fixtures (yet,
probably). For this reason I subclass the standard test runner and then
force it to do nothing to the databases. This should make it possible to
easily run tests against the database.