Pass all tests
This now handles all cases (excluding the one expected failure). The
failure is because of uncertainty over the rna_type, which will be
resolved once more databases are installed.

I have had to change a few tests to match the descriptions but overall
we now pass all tests and show the expected behaviors. However, the code
is very slow due to far too many database queries. This needs to be
cleaned up and fixed.