Make better use of django ORM for speed
The ORM can be much faster than it was before. This reduces the run time
of the largest case I test (10k xrefs) to ~3 seconds. Previously it
could take a long time (minutes) to compute the description. This was
due to inefficient usage of the ORM where it basically had to select
each xref one at a time, leading to many queries. We now do a better job
of using the ORM and querying to be much faster.

This is a large improvement over the previous method. While I have
tested raw sql and get ~1 second for run time, it is not worth the
addition of raw sql strings into our code base.

To make this possible I refactored the get_description/get_rna_type
functions slightly to take the list of xrefs as input and moved the
valid_xref method into the Rna model. This version should cover all the
basic functionality we need to improve naming.