Improve query performance
The query has been rewritten to get all xrefs, rna, and accessions in
one go. This should be much faster overall. In looking at the output I
saw that the previous version could pause for a long time between
writing results. That happened because for each rna entry we fetched
all xrefs, leading to the pause. This is no longer the case as the query
is now based upon xrefs and can get all of them, as well as group them
in one go.

The grouping is done in python lazily, and hopefully that will not be
too much of a penalty.