Correct implementation
This should now produce data for all sequences, instead of just those
with 'good' xrefs. In addition, I've moved to using Serializers from
rest_framework to be more consistent with the rest of the code base. As
part of that I'm using namedtuples as an intermediate stage between the
model and the Serializers because there is some data that does not
cleanly fit into the model. Using attrs classes would be better, but
that isn't needed yet.

Finally, I've modified the get_data method to make testing this a bit
easier.