Add min and max options
Trying to export without a min/max is probably glacially slow. I tried
and the process got killed after 14 hours. In order to make the
exporting work in a reasonable time frame I've added the min/max
options. These work the same way min and max do in the xml export and
are required. However, there is an important caveat with this sort of
export.

The min/max constraints apply to the number of entries in the RNA table
are extracted. Leading to the fact that the number of entries in the
produced JSON file may be less than expected from the given range. For
example, if asked for 1 to 1000 you will get ~700 entries. This is due
to the constraints on xref and accession where the xref cannot be
deleted and the accession must have an external_id.

Without those constraints on xref and accession, it is possible to get
more output than the expected range. I wouldn't be surprised to learn
that there is a set of min/max cutoffs that show the same behavior
(produce more sequence entries than expected from the range).

It is also an error to specify both --test and min/max.