Add a computed gene name
There are several examples of genes that do not have the expected gene
names. For example, dme-bantam-5p does not have the gene name starting
with bantam. This is a problem because our search logic will not split
on '-' (to prevent extra matches). But this means searching for 'bantam'
does not find all entries with the word 'bantam' in them. To deal with
this I change the logic of creating the gene names to strip the leading
three letters and -. This is a common pattern in miRBase. Doing so
should improve our search as we will now find the expected hits for
bantam. We shouldn't hurt anything as well because searching for
dme-bantam will still find the same things as the product name will
still contain it.