Do not export given rna_type, use computed one
Previously we would use the given (from each database) RNA type for
exporting in the XML file. This has some downsides where often the given
RNA type is not what we think is 'correct'. For example, some databases
will call something that is a precursor as a miRNA. I have added logic
to correctly distinguish these cases and store the improved (hopefully)
RNA type in previous commits.

It now makes sense to use the computed RNA type in export so that when
the data indexed and searched users see our improved annotations and not
the given ones. As a concrete example, we no should no longer see
precursors when listing out all miRNA's.

A potential downside to this change is that people may have become
reliant upon the other databases behavior. Hopefully a generous enough
search will get around this though. They should be able to search by
name of the sequence and still find it. In addition, we are assuming
that we can do a better job of selecting the RNA type than other
databases.

In addition, exporting XML is now implicitly dependent upon the
precomputation steps. Without precomputing many sequences will not be
indexed.