Allow export of multiple RNA types for antisense
After examining what would happen for some cases I can see that we would
no longer call several things which are lncRNA's as lncRNA. This would
have the effect of not finding all lncRNA's when browsing. As an example
consider:

http://test.rnacentral.org/rna/URS0000732D5D/9606

This has a precomputed RNA type of 'antisense' (likely because if we
cannot determine the RNA type then we use majority vote then
alphabetical ordering and antisense < lncRNA. Because of this we would
not index it as an lncRNA and so it could not be found using the facets.
This is not ideal as it is an lncRNA and should be listed as such. The
current logic will turn the antisense facet into a grab bag of stuff
that is labeled antisense and something else.

To deal with this if we have precomputed and RNA type as antisense then
we will also allow all other annotations. This means the above example
would not be lost (which is good) but does not improve our searches in
some cases. For example if we have something that is precomputed as
antisense, but has been given as miRNA and lncRNA it will be found under
all 3 facets. In reality it can't be a miRNA, ~22 nt long, and an
lncRNA, > 100 nt long, (though it could be an lncRNA containing an miRNA
that is antisense to a protein).

This however is the safest option. Especially since we don't know who
labels something antisense and why. Until we do, we cannot safely deal
with it. Once we learn more we will revisit this change as well as how
to correctly compute the RNA type in this case.