Deal with no computed rna_type
There should always be a computed RNA type but it is possible for there
to be none. We should allow this to happen and deal with. What we
basically do is fall back to the previous behavior of using all given
rna_types. This is safest won't alter the results. It would be better to
warn or something when this happens though.