Improve ensembl export
This improves the logic of exporting for ensembl in several ways. First,
we will now export all sequences that we would display (all those with
active xrefs). Secondly we have a better choice of databases to display
xrefs for. We also speed up the process quite a bit. A test run of 10000
can now be done in minutes instead of hours (thanks boris!).

Also some small bug fixes are included.