Shift coordinates for BED files #123
BED files are 0 based, not 1 based unlike what is stored in our
database. This should fix the issue so that this will write 0 based BED
files instead of 1 based. However, it is possible that some coordinates
are already 0 based and we don't know it. This would produce the wrong
coordinates in that case. I don't have a good way to test this other
than to use bedtools to generate all sequences and then test the MD5
hashes for them. That would be very time consuming so I'm not planning
on doing it.