Use same xrefs for gff and bed (#124)
This alters bed file export for Rna objects so that it will use the same
xrefs that GFF file export would. I'm not sure why they differed, and
their could be a reason. However, it is very surprising that there is a
difference and so I'm eliminating it in favor of the logic in GFF
export. The reason being there are xrefs missing in GFF in BED but there
are no BED only entries for pombe.

It would probably be better to pull out the xref logic for these two
methods (and probably several others) so that the same xrefs are used
consistently over the code. That said I'm not sure if there is a single
correct set of xrefs or how to identify it.