Initial addition of a JSON schema
A useful idea is to validate our export with a JSON schema
(https://github.com/RNAcentral/rnacentral-webcode/pull/128). This is my
first attempt at creating such a schema. This is based off me looking
at the example object Anton provided in the issue. This hasn't been
tested in anyway yet, but is just my thoughts on what the schema should
contain.

One issue I am worried about is being overly strict. I went as far as
adding enum's for all database we currently export plus all possible
rna_types. I'm not sure this is a good idea or if it will cause a lot of
pain when exporting.