Initial addition of a mapping dict
The goal here is create a mapping from INSDC terms to correct SO term as
seen in https://github.com/RNAcentral/rnacentral-webcode/issues/63. This
is my initial start at this. I don't think I will just leave a random
dict in some file, but I need to put this somewhere for now. Future work
will fill out this dict more as well as move it around (probably). For
now this is just to fill out the mapping.