Update mapping to use SO terms
This is hopefully a better assignment to SO terms (instead of GO or
other terms). Most of the assignments are very straightforward. There
are a few (3) assignments that aren't so I have provided some notes on
why the SO term was selected.

In addition there is an initial implementation of the function to
determine the SO term for the given URS. The function is very simple and
probably should include some logging to indicate when it does not have a
good mapping to SO terms as well.