Use passed validator and ignore bad data
It is possible that entires in the database may not be complete (empty
rna_type) or no precomputed data is present. These are likely due to an
incomplete import process. However we should still deal with these
errors. Right now it silently ignores them which is not ideal but this
at least allows the process to continue.

In addition, I've implemented the uniqueness constraint on the Xref's
produced. This required adding some dunder methods and placing them into
a set. The sorted call is used to reliably produce the same output each
run.