Change precursor_RNA to map to primary_transcript
As discussed in #63 the mapping to pre-miRNA is too precise unless we
have additional information. At the moment the mapping is just based
upon the terms and does not consider any additional information. Because
of this I use a more general term for now. We can use a more precise one
once the mappings consider other things. For example, if the data is
coming from miRBase it is probably a pri-mRNA, however we don't do that
yet so we will stick with a more generic mapping.