Initial work on extending the indexed fields
I'm adding standard_name and locus_tag as those are often useful fields
to search with. For example from Ensembl/GENCODE and Wormbase we get
some really useful information out of them.

I'm not sure this actually works correctly yet. I've added the fields to
the query and to the exported redundant fields. I haven't checked to see
if they need to be escaped in anyway yet.