Add GENCODE/Ensembl to description logic
This adds GENCODE and Ensembl to the list of databases to select a name
and RNA type from. I place GENCODE above HGNC as it is also manually
curated but we really like it. Ensembl is always better than one of the
generic database so I put it further down but not last. Maybe it should
be moved higher than it is now but we will see what happens with this
version.