Initial untested design
This is my initial design of how to represent Rfam models. This is may
be too complex for our usecase, but this is my initial approach. My goal
was to create a comprehensive normalized representation of the search
results, as I think this will make analysis and display easier later.

One notable thing this does not deal with is the versioning and changes
that may occur in Rfam. For example we do not assign Rfam families to
versions, nor do we track changes to Rfam families. We should to be
complete and make reproducibility easier but I'm unsure of how essential
that is to our work.