Simplify models as suggested
After talking with @AntonPetrov about the design of the tables to store
rfam information we decided to simplify the design of hits and add some
extra tables.

The additional tables are needed to keep track of what families each
sequence is part of, as well as which sequences were searched.