Use psycopg2 instead of orcale
Moving to postgres instead of Oracle, so I'm replacing the dependency on
oracle. In addition, I'm stripping it from the Dockerfile and the env
template. I have also altered the local_settings_default.py file to show
the new fields that have to be filled out to connect, and removed the
unsupported 'threaded' option.

Note, that this probably does not handle the exporting scripts as some
of those scripts refer directly to cx_oracle and oracle specific things.
That can be fixed later as this is enough to start working with.