Improve display of Rfam matches
This is some work to make the display of Rfam hits better. The idea is
to make it clearer and easier to know if the results are correct or
incorrect. We want to show a simple checkbox if the result is good
(correct matches, or no match but that is expected) vs when there is a
possible issue (incorrect matches, no matches but there should be). This
does not deal with the complex case of 3 matches and 2 are good but one
is not yet. Also we do not show anything if there is no analysis for the
given sequence yet.