Implement simple display for rfam matches
This now shows the result of using Rfam models on a sequence. The
display is fairly simple now as it will not deal with cases like 2
families match and 1 is correct and a second is not. It does however,
provide some simple context about the match. For example, we add
tooltips to state if the match(es) are correct (no match but that is
expected, or the matches have the correct rna_type), or incorrect (match
when there should be none, or a match that has the wrong rna_type).

In addition, we 'deduplicate' the matches before display by selecting
only the model with the lowest e-value and highest score for each clan
for display. This prevents issues where a single sequence matches
several families in a clan (eg rRNA's).

I changed the rna_type that is displayed for each sequence to use the
get_rna_type() method. My reason being that that method will compute the
rna_type if needed. And as the testing database is out of sync now we
would otherwise get a lot of 'None' values. This prevents that.