Simplify rfam migrations
There were two migrations before and now there is only a single simpler
one. Since I changed the table names I didn't want to add another
migration or alter the existing ones (as I would probably make a
mistake). So I deleted the old migrations and created a new migration
instead. This would not be good if anyone applied it yet, but I doubt
that is the case.