Add docstrings
This should help make the code more understandable. Hopefully this is
enough to describe what the code currently does. Note that the
get_rfam_hits method will change in response to another code review
comment but for now the docstring describes what it is doing, not what
it should do.