Attempt to implement rfam family name export
This is my first attempt at adding rfam family name export. It will be
useful to search by rfam family name. I'm not sure if the name we export
with is required to be the same as the search name and if so we should
probably shorten this to a different name. This assumes that there may
be more than one rfam family, which is true for large sequences, and
exports them like we export rna_types.