Implement finding and display of sequence fragments
To begin with I refactored the display and logic for finding any Rfam
issues. This was driven by the fact there there are at least 3 types of
issues we want to display:

1. The sequence and model are from different domains
2. The sequence is just a fragment of a larger sequence
3. The sequence and model have conflicting RNA types

Because this will work out to a fair amount of code, I've moved that
logic into it's own module (portal.rfam_matches) and provided some
objects to manage the logic as well as the results. The general design
is that we have a single function (check_issues) which checks all the
objects that determine if there is any issue. The issues are stored in
the RfamMatchStatus object.

This object is then used to display each message in the template so the
user can see any issue(s) that exist. Currently, the display will not do
a good job with sequences that have more than one issue as the message
are long and they will all be placed on one line.

The only currently implemented detection is that of sequence fragments.
It is very useful to detect sequences that are fragments of a larger
sequence and this can be done by seeing if the sequence is nearly
entirely matched by a model but the model is much larger than the match.