Implement domain issue detection
This can find if a model and sequence have differing domains. Currently
this is pretty simple in that it requires there be only a single hit to
the sequence and that the sequence have a single domain. This may not
always be ideal but it does work on the given example of URS00002EF971.