Add a column for storing precomputed rfam data
It is useful for indexing and the like to have a column of precomputed
data about the issues detected from the Rfam scan. The column will be a
simple JSON object but the version of Django we are using does not
support postgres's JSON column type. Maybe once we upgrade we can use
such a datatype.