Initial addition of rfam indexing
This is the first attempt at producing xml files that can be indexed to
track the Rfam match information. With this method we will allow users
to search for sequences which match a particular family, either by name
or id, as well as clan, and search for issues. Notably we can make all
searches find only the sequences that appear to have no issues, which
will be a very important quality check for users.