Attempt to exclude unclassified domains
We should not care if an unclassified organism matches a sequence with a
known domain. This is a good thing actually. Thus I special case the
domain logic so we can exclude any domains for an environmental sample or
uncultured organisms. This should prevent us from seeing many
uninformative warnings.