First attempt at showing Rfam matches
We want to show all the rfam hits in a sequence in it's own section.
This section is below where the genoverse viewer should appear. If no
hits a present then a simple 'No hits' message appears. This probably
can be refined more.

In the case where there are hits we show a fairly structured section
that summarizes some important information. We show the 2D of the
family, which could confuse some users into thinking it is the 2D of the
sequence. Some testing and work is needed to clean that up. We also
state where the match occurs and how much of the model matches. Later we
will use a sequence viewer to show such features, but for now this will
do.

At the moment the GO terms that are shown are hardcoded. This is because
we do not store GO terms yet. Next will be to store such terms and use
them here.

In addition, the rna_type label in the overview is now a badge. The hope
is that making this a badge, as well as the label in the rfam hits
section, will make the user compare the two. I don't use different
colors between them and I'm not sure if I will quite yet.