Create tables to store GO terms
This adds a simple table to store GO term information (GO term id and a
name) as well as a table to map from Rfam family to GO terms. The GO
term ids are like 'GO:0003735'. The design is very simple now, just
enough to show a little context and link out. Future work may need a
more complex design but this is enough for Rfam hits.