Add ability to export specific UPIs
In order to see what the issue is with exporting multiple RNA types I
need to be able to export metadata specific sequences. While it is
possible to lookup the id and the like it is easier and quicker to just
use the UPI directly. Thus I have added this feature. The xml_export now
accepts either a --upis with a comma separated list of upis to export or
a --min and --max option as before.

As part of this the code underwent a major refactoring. I couldn't
really follow the logic with deeply nested functions. This way the code
is written more or less in the order it is executed. As part of the
refactoring I fixed a small bug where if gzip failed the program did not
fail.