Prevent adding multiple RNA types (#161)
Previously this would and another RNA type if the RNA type was
antisense. On one hand this is kind of useful because antisense
generally more useful with another RNA type. However, this leads to
confusing search results where some RNA type facets are not unique.

This is fixed so there can only ever be one RNA type. This will change
the RNA in search. For example URS00007703DB will now be 'antisense RNA'
instead of 'lncRNA'. This change may not be ideal but I'd rather work on
getting the RNA type detection logic to call that sequence lncRNA
instead of adding another annotation.