Add table to store secondary structures
As part of the GtRNAdb import we will be storing secondary structures.
This requires us to have a table to store the pairing information into
the database. This is a very simple table that will store the secondary
structure pairing information as a dot-bracket string. The secondary
structure is linked to an accession as the information comes from some
database. We allow for a single accession to have more than one
secondary structure, as this is possible. This does not contain any
information on the layout of the secondary structure (how to draw it) as
for now we will stick to something like forna to draw it.