Detect if missing Rfam hit
This is an initial implementation of detecting if a sequence is missing
a match it should have. Some sequences should be labeled as tRNA by Rfam
but do not have a match to the tRNA family. This is the first attempt at
it and likely isn't the final solution. Currently this is limited to
tRNA and rRNA only.