First attempt at correcting HGNC descriptions
HGNC had some useful suggestions on changes the naming scheme they want.
First, they don't want us to select high gene numbers (which was an
accident) like RNA5S17 instead of RNA5S1. This should be done by
selecting descriptions from accessions with the smallest gene.

Also, for URS that represent more than one gene they want an indication
of that. So I've added a list of the genes if there are few (< 3) and
otherwise the text 'multiple genes' will appear. Untested, so far but
this is the general approach.

I've also altered the default behavior to bias it away from larger gene
numbers and toward smaller ones, hopefully.