Improve HGNC and mirBase naming
In both databases it is possible for more than one gene/precursor to map
to a single URS. In this case we need to use a better naming system than
the current one. Notably we should indicate if there are several items.
This means listing the genes/precursors or a note to say there is more
than one. This adds some logic to do so. It includes logic to do things
like correctly sort names (basically stripping trailing numbers if
possible) so things are listed in the right order. The code is rather
obscure at the moment but can be corrected.