Read rfam data and export tax_string
The tax_string is the classification and provides a seach like Rfam does
for taxonomy. This should make finding specific organisms or groups
easier.

Also, I've added the logic for querying the Rfam hits back in. This
required rewriting the query to use explicit joins because we must do
outer joins and we can't seem to mix references to those implicitly
joiend tables in outer joins.