Expand the Rfam classification when possible
If we have a single Rfam match then the match should be allowed to take
up all the room. This is nice in cases where the GO term description is
long as this way it will not split into several lines unnecessarily. This
adds that functionality. Two examples are:

http://localhost:8000/rna/URS000080E226/274

and

http://localhost:8000/rna/URS00000478B7/9606

In the second there is a large space between the image and the text but
that is because that image is very narrow. Shrinking the space to the
text in that case makes the first case with a wide image look bad.

I've also added a with statement for the grouped hits as I end up using
it several times when detecting the classes.