Fix issue with overlapping text/image
This seems to prevent the issue of the text and Rfam image from
overlapping. Basically I increased the size of the section for the rfam
image in a medium sized window. I also added a setting if it is in a
large window so the text and image will not be too far away in a large
window.