Improve HGNC -> ensembl mapping
This adds the ability to generate more mappings. If ensembl has only a
single transcript we use that. In cases where there is more than one we
pick the longest. I retain the sequence based approach to not accidentally
change anything, but I think it should work the same as mapping to the
longest.

I also added some logic to fail mapping if the ensembl API shows issues.
This may be a bad idea.