Only apply rna type stripping to RefSeq
We don't want to apply the RefSeq RNA type stripping to any database
other than RefSeq right now. This can be a problem with Rfam as Rfam
will use the RNA type in the generated names, so stripping it leads to
really bad names. Thus I will move to be very conservative about this
for now.