Do not allow tRNA's to be more than 700nt
GtRNAdb does not have tRNA's that are this large. The tRNA's that are
larger than 700nt are generally something strange. They often contain a
tRNA within it but should not be labeled as a tRNA. For example:

- http://rnacentral.org/rna/URS0000A17B82/640938
- http://rnacentral.org/rna/URS000083F182/242161

Neither of those are a tRNA, while they both contain tRNA's.

This doesn't cover all the cases yet. For example:

- http://rnacentral.org/rna/URS00005A245E/10090

This probably shouldn't be labeled as a tRNA but will be under this
system. This will only alter 75 sequences right now. I will work to
improve it more.