Prefer lncRNA over antisense_RNA
In some cases a sequence is labeled as both antisense_RNA and lncRNA. We
will prefer lncRNA over antisense_RNA now. Generally what people mean in
these cases is 'long non-coding RNA that is antisense to a protein
coding gene'. I think it is clearer to just use lncRNA. With the move to
SO terms we should be able to express this case more clearly.