Determine when sequences are mitochondrial
In some cases (25 to be exact) HGNC can tell us that a molecule is
stored in the mitochondria. In those we store that information. Most of
the location entry is stuff like `7q32.1`, which is a position on a
nuclear chromosome. This is a minor improvement but does make the HGNC
annotations consistent with other databases in these cases, as
previously all other databases would state the molecule is found in the
mitochondria.