Update tests and naming logic
This updates the naming logic and tests. We add more tests to ensure
that some issues/sequences that were shown in the SAB will always have
the correct naming. Additionally, I've changed some logic to better
handle certain edge cases. For example if we have a sequence named
'predicted', we now try to find an Rfam sequence and then use that
sequence along with any gene annotation. Hopefully this improves the
naming of other cases.