Remove extra Rfam based naming
This prevents the code from using rfam as a source of names in cases
where Rfam may have a better name than the selected database. The idea
was to use an Rfam name if the name had 'predicted' in it. However, this
lead to problems where Rfam and the database disagreed on what RNA type
the sequence is. This is fixable, but I am skipping it for now.