Show a loading icon while the image is loading
This adds a loading icon while the chart is being fetched. This works by
changing the fetch to get base64 data and then use that as the src. This
is needed because previously doing two requests gave us errors for some
reason. This isn't perfect because initially the modal has an invalid
URL that is fetched.