Remove rfam_problems_field
This field is no longer populated by the pipeline or used by the
webcode. It is now repalced by entries in the qa_status table so it
isn't needed. I am removing it as part of a general cleanup of the
precompute table, as it isn't a good idea to keep obsoleted code around.