Add angular component to show QA status
This adds an angular component which will display the QA status on
sequence pages. The previous method passed some QA information along
with the Rfam hits and displayed that. The information on the Rfam hits
was computed by django and not the information stored in the database
for some reason. I'm unclear as to why this was done.

Instead what this does is use the messages that are stored in the
database and thus provided by the API and displays those in a new
section in the webpage. This QA section is near the sequence viewer so
it isn't too much of a change from the current website but likely could
be better designed. This also assumes that the messages in the table are
always correct html, which probably isn't ideal.