Add ORF to feature viewer
This modifies the API to include CPAT ORFs and then modifies the feature
viewer to show them as well. The viewer changes include removing a
the duplicated code and factoring out some of the common setup steps.
With these changes it is more like configuration to build a new entry in
the feature viewer. More could be done but this is simple enough.