Initial addition of poetry/pre-commit
This is the initial version of pre-commit and poetry. Note that there is
no poetry lock or check hooks though there should be. This is because
those files are not at the top level of the project. I think they should
be moved but I don't want to break our entire setup yet.