Update docs to include database dumps
We will start producing pg_dump files as part of the release so we
should document them. This is an initial outline of where they are kept.
Maybe we should include instructions on how to use them, but I don't
want this to appear too easy to use. Ideally, most people use our
database (so we can track and see we are heavily used), but a few power
users download the database dumps.