Release 21 (#602)
* Add logos for expression atlas and plncdb

* Update release version in the badge

* Add data about release 21

* Use release-v21 image

* Update RELEASE_ANNOUNCEMENT_URL

* Add release-v21 image

* Update version and add Expression Atlas

* Rename image

* Update the number of articles found

* Fix endpoint used by LitScan

* Update text about the query used

* Update docs to include database dumps

We will start producing pg_dump files as part of the release so we
should document them. This is an initial outline of where they are kept.
Maybe we should include instructions on how to use them, but I don't
want this to appear too easy to use. Ideally, most people use our
database (so we can track and see we are heavily used), but a few power
users download the database dumps.

* Fix annotation link for Expression Atlas

* Add expression_atlas_id field

* Add link to experiments

* Remove expression_atlas_id field from the API

Co-authored-by: Blake Sweeney <bsweeney@ebi.ac.uk>